Move soil structs to separate module
