+++ SharedHydrology/Project.toml
@@ -1,0 +1,11 @@
+name = "SharedHydrology"
+uuid = "e4c7f7b8-5a3d-4e6f-9c1a-2b8d3f4e5a6c"
+version = "0.1.0-dev"
+authors = ["Deltares and contributors"]
+
+[deps]
+StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
+
+[compat]
+julia = "1.10"
+StaticArrays = "0.12, 1.0"
